refactor: extract SceneFieldCodeGenerator delegate (1112→907 lines)
Move 15 field code-generation methods from StorytellingSceneEditor into
a SceneFieldCodeGenerator delegate class:

- getCurrentStateCodeForField, generateCodeForSetUp
- getDoStatementsFor{CopyField,AddField,RemoveField}
- getUndoStatementsFor{AddField,RemoveField}
- getRiders plus 7 private helpers (replaceReferencesInExpression,
  fillInAutomaticSetUpMethod, doesSetVehicleImplyVehicle, etc.)

Override methods in SSE become thin wrappers. Static utilities
(asSetVehicleCall, isSetVehicleInvocation) remain accessible to
useSceneAsVehicleForDisconnectedModels.

Adds 20 contract tests in SceneFieldCodeGeneratorTest.
Removes 6 now-unused imports from SSE.

Part of #528.
