+++ core/ide/src/main/java/org/alice/stageide/sceneeditor/StorytellingSceneEditor.java
@@ -45,11 +45,8 @@ package org.alice.stageide.sceneeditor;
 import com.formdev.flatlaf.extras.FlatSVGIcon;
 import edu.cmu.cs.dennisc.animation.Animator;
 import edu.cmu.cs.dennisc.animation.ClockBasedAnimator;
-import edu.cmu.cs.dennisc.java.lang.ArrayUtilities;
 import edu.cmu.cs.dennisc.java.lang.SystemUtilities;
-import edu.cmu.cs.dennisc.java.util.Lists;
 import edu.cmu.cs.dennisc.java.util.logging.Logger;
-import edu.cmu.cs.dennisc.pattern.IsInstanceCrawler;
 import edu.cmu.cs.dennisc.render.OnscreenRenderTarget;
 import edu.cmu.cs.dennisc.render.RenderCapabilities;
 import edu.cmu.cs.dennisc.render.event.*;
@@ -79,7 +76,6 @@ import org.alice.math.immutable.*;
 import org.alice.nonfree.NebulousIde;
 import org.alice.stageide.StageIDE;
 import org.alice.stageide.croquet.models.sceneditor.ViewListSelectionState;
-import org.alice.stageide.modelresource.ClassResourceKey;
 import org.alice.stageide.oneshot.DynamicOneShotMenuModel;
 import org.alice.stageide.run.RunComposite;
 import org.alice.stageide.sceneeditor.interact.CameraNavigatorWidget;
@@ -100,8 +96,7 @@ import org.lgna.project.ast.*;
 import org.lgna.project.virtualmachine.UserInstance;
 import org.lgna.story.*;
 import org.lgna.story.implementation.*;
-import org.lgna.story.implementation.alice.AliceResourceClassUtilities;
-import org.lgna.story.resources.ModelResource;
+
 
 import javax.swing.Icon;
 import javax.swing.SwingUtilities;
@@ -109,7 +104,6 @@ import java.awt.Dimension;
 import java.awt.Graphics;
 import java.awt.event.MouseEvent;
 import java.util.*;
-import java.util.List;
 
 /**
  * @author dculyba
@@ -128,6 +122,7 @@ public class StorytellingSceneEditor extends AbstractSceneEditor implements Rend
   }
 
   private final SceneEditorDropReceptor dropReceptor = new SceneEditorDropReceptor(this);
+  private final SceneFieldCodeGenerator codeGenerator = new SceneFieldCodeGenerator(this);
 
   private StorytellingSceneEditor() {
   }
@@ -685,7 +680,7 @@ public class StorytellingSceneEditor extends AbstractSceneEditor implements Rend
 
   private void useSceneAsVehicleForDisconnectedModels(UserMethod generatedSetupMethod) {
     for (Statement statement : generatedSetupMethod.body.getValue().statements.getValue()) {
-      MethodInvocation setVehicleCall = asSetVehicleCall(statement);
+      MethodInvocation setVehicleCall = SceneFieldCodeGenerator.asSetVehicleCall(statement);
       if (setVehicleCall == null) {
         continue;
       }
@@ -694,23 +689,6 @@ public class StorytellingSceneEditor extends AbstractSceneEditor implements Rend
         args.getFirst().expression.setValue(new ThisExpression());
       }
     }
-  }
-
-  private MethodInvocation asSetVehicleCall(Statement statement) {
-    if (statement instanceof ExpressionStatement expressionStatement) {
-      Expression expression = expressionStatement.expression.getValue();
-      if (expression instanceof MethodInvocation mi) {
-        Method method = mi.method.getValue();
-        if (method.getName().equalsIgnoreCase("setVehicle")) {
-          return mi;
-        }
-      }
-    }
-    return null;
-  }
-
-  private boolean isSetVehicleInvocation(Statement statement) {
-    return asSetVehicleCall(statement) != null;
   }
 
   @Override
@@ -737,10 +715,6 @@ public class StorytellingSceneEditor extends AbstractSceneEditor implements Rend
     this.globalDragAdapter.setHandleVisibility(true);
   }
 
-  private void fillInAutomaticSetUpMethod(StatementListProperty bodyStatementsProperty, boolean isThis, AbstractField field, boolean getFullFieldState) {
-    SetUpMethodGenerator.fillInAutomaticSetUpMethod(bodyStatementsProperty, isThis, field, this.getInstanceInJavaVMForField(field), this.getActiveSceneInstance(), getFullFieldState);
-  }
-
   @Override
   public void setFieldToState(UserField field, Statement... statements) {
     EntityImp fieldImp = getImplementation(field);
@@ -753,220 +727,41 @@ public class StorytellingSceneEditor extends AbstractSceneEditor implements Rend
 
   @Override
   public Statement getCurrentStateCodeForField(UserField field) {
-    Statement rv = null;
-    BlockStatement bs = new BlockStatement();
-    fillInAutomaticSetUpMethod(bs.statements, false, field, true);
-
-    Statement setVehicleStatement = null;
-    for (Statement statement : bs.statements.getValue()) {
-      if (isSetVehicleInvocation(statement)) {
-        setVehicleStatement = statement;
-        break;
-      }
-    }
-    if (setVehicleStatement != null) {
-      bs.statements.getValue().remove(setVehicleStatement);
-    }
-    DoTogether dt = new DoTogether(bs);
-    if (setVehicleStatement != null) {
-      DoInOrder dio = new DoInOrder(new BlockStatement(setVehicleStatement, dt));
-      rv = dio;
-    } else {
-      rv = dt;
-    }
-
-    return rv;
+    return codeGenerator.getCurrentStateCodeForField(field);
   }
 
   @Override
   public void generateCodeForSetUp(StatementListProperty bodyStatementsProperty) {
-    AbstractField sceneField = this.getActiveSceneField();
-    this.fillInAutomaticSetUpMethod(bodyStatementsProperty, true, sceneField, false);
-    for (UserField userField : this.getActiveSceneType().getDeclaredFields()) {
-      if (userField != null) {
-        if (userField.getManagementLevel() == ManagementLevel.MANAGED) {
-          this.fillInAutomaticSetUpMethod(bodyStatementsProperty, false, userField, false);
-        }
-      }
-    }
-  }
-
-  private Statement replaceReferencesInExpression(UserField fieldToReplace, UserField replacement, Statement statement) {
-    IsInstanceCrawler<FieldAccess> crawler = IsInstanceCrawler.createInstance(FieldAccess.class);
-    statement.crawl(crawler, CrawlPolicy.COMPLETE, null);
-
-    for (FieldAccess fieldAccess : crawler.getList()) {
-      AbstractField field = fieldAccess.field.getValue();
-      if (field == fieldToReplace) {
-        fieldAccess.field.setValue(replacement);
-      }
-    }
-    return statement;
+    codeGenerator.generateCodeForSetUp(bodyStatementsProperty);
   }
 
   @Override
   public Statement[] getDoStatementsForCopyField(UserField fieldToCopy, UserField newField, AffineMatrix4x4 initialTransform) {
-    Statement stateCodeStatement = this.getCurrentStateCodeForField(fieldToCopy);
-    stateCodeStatement = replaceReferencesInExpression(fieldToCopy, newField, stateCodeStatement);
-
-    //Remove the setVehicle and setTransform statements from the setup code, so we can replace them with custom ones based on the fieldToCopy's vehicle and the initial transform
-    List<BlockStatement> blockStatements = new LinkedList<BlockStatement>();
-    if (stateCodeStatement instanceof BlockStatement statement) {
-      blockStatements.add(statement);
-    } else if (stateCodeStatement instanceof AbstractStatementWithBody body) {
-      blockStatements.add(body.body.getValue());
-    }
-    while (!blockStatements.isEmpty()) {
-      BlockStatement bs = blockStatements.removeFirst();
-      Statement setVehicleStatement = null;
-      Statement setPositionStatement = null;
-      Statement setOrientationStatement = null;
-      for (Statement s : bs.statements.getValue()) {
-        if (s instanceof BlockStatement block) {
-          blockStatements.add(block);
-        } else if (s instanceof AbstractStatementWithBody body) {
-          blockStatements.add(body.body.getValue());
-        } else if (s instanceof ExpressionStatement expressionStatement) {
-          Expression expression = expressionStatement.expression.getValue();
-          if (expression instanceof MethodInvocation mi) {
-            Method method = mi.method.getValue();
-            //Look for the setVehicle, setOrientation, and setPositions for the field. Note that we need to make sure these calls are being called on the field and not the joints, hence the check for FieldAccess (joints are called off of getJoint and resolve as a MethodInvocation)
-            if (method.getName().equalsIgnoreCase("setVehicle") && (mi.expression.getValue() instanceof FieldAccess)) {
-              setVehicleStatement = s;
-            } else if (method.getName().equalsIgnoreCase("setOrientationRelativeToVehicle") && (mi.expression.getValue() instanceof FieldAccess)) {
-              setOrientationStatement = s;
-            } else if (method.getName().equalsIgnoreCase("setPositionRelativeToVehicle") && (mi.expression.getValue() instanceof FieldAccess)) {
-              setPositionStatement = s;
-            }
-          }
-        }
-      }
-      if (setVehicleStatement != null) {
-        bs.statements.getValue().remove(setVehicleStatement);
-      }
-      if (setPositionStatement != null) {
-        bs.statements.getValue().remove(setPositionStatement);
-      }
-      if (setOrientationStatement != null) {
-        bs.statements.getValue().remove(setOrientationStatement);
-      }
-    }
-
-    Object toCopyInstance = this.getInstanceInJavaVMForField(fieldToCopy);
-    AbstractField toCopyVehicleField = null;
-    if (toCopyInstance instanceof Rider rider) {
-      SThing vehicleInstance = rider.getVehicle();
-      toCopyVehicleField = this.getFieldForInstanceInJavaVM(vehicleInstance);
-    }
-    Statement[] initializeStatements = SetUpMethodGenerator.getSetupStatementsForField(false, newField, this.getActiveSceneInstance(), toCopyVehicleField, initialTransform);
-    Statement[] statementsToReturn = new Statement[initializeStatements.length + 1];
-    System.arraycopy(initializeStatements, 0, statementsToReturn, 0, initializeStatements.length);
-    statementsToReturn[initializeStatements.length] = stateCodeStatement;
-    return statementsToReturn;
+    return codeGenerator.getDoStatementsForCopyField(fieldToCopy, newField, initialTransform);
   }
 
   @Override
   public Statement[] getDoStatementsForAddField(UserField field, AffineMatrix4x4 initialTransform) {
-    if ((initialTransform == null) && field.getValueType().isAssignableTo(SModel.class)) {
-      AbstractType<?, ?, ?> type = field.getValueType();
-      JavaType javaType = type.getFirstEncounteredJavaType();
-      Class<?> cls = javaType.getClassReflectionProxy().getReification();
-      Class<? extends ModelResource> resourceCls = null;
-      if (SModel.class.isAssignableFrom(cls)) {
-        resourceCls = AliceResourceClassUtilities.getResourceClassForModelClass((Class<? extends SModel>) cls);
-      }
-      Point3 location;
-      if (resourceCls != null) {
-        ClassResourceKey childKey = new ClassResourceKey((Class<? extends ModelResource>) cls);
-        AxisAlignedBox box = childKey.getBoundingBox();
-        boolean shouldPlaceOnGround = childKey.getPlaceOnGround();
-        double y = (box != null) && shouldPlaceOnGround ? -box.getXMinimum() : 0;
-        location = new Point3(0, y, 0);
-      } else {
-        location = Point3.ORIGIN;
-      }
-
-      initialTransform = new AffineMatrix4x4(OrthogonalMatrix3x3.IDENTITY, location);
-    }
-    return SetUpMethodGenerator.getSetupStatementsForField(false, field, this.getActiveSceneInstance(), null, initialTransform);
+    return codeGenerator.getDoStatementsForAddField(field, initialTransform);
   }
 
   @Override
   public Statement[] getUndoStatementsForAddField(UserField field) {
-    List<Statement> undoStatements = Lists.newLinkedList();
-
-    undoStatements.add(SetUpMethodGenerator.createSetVehicleNullStatement(field));
-
-    return ArrayUtilities.createArray(undoStatements, Statement.class);
+    return codeGenerator.getUndoStatementsForAddField(field);
   }
 
   public Map<AbstractField, Statement> getRiders(UserField vehicle) {
-    IDE.getActiveInstance().ensureProjectCodeUpToDate();
-    UserInstance sceneAliceInstance = getActiveSceneInstance();
-    UserMethod generatedSetupMethod = sceneAliceInstance.getType().getDeclaredMethod(StageIDE.PERFORM_GENERATED_SET_UP_METHOD_NAME);
-    Map<AbstractField, Statement> riders = new HashMap<>();
-    for (Statement statement : generatedSetupMethod.body.getValue().statements.getValue()) {
-      MethodInvocation setVehicleCall = asSetVehicleCall(statement);
-      if (setVehicleCall != null && doesSetVehicleImplyVehicle(setVehicleCall, vehicle)) {
-        FieldAccess riderAccess = (FieldAccess) setVehicleCall.expression.getValue();
-        riders.put(riderAccess.field.getValue(), statement);
-      }
-    }
-    return riders;
-  }
-
-  private boolean doesSetVehicleImplyVehicle(MethodInvocation setVehicleCall, UserField vehicle) {
-    ArrayList<SimpleArgument> args = setVehicleCall.requiredArguments.getValue();
-    if (args.size() == 1 && setVehicleCall.expression.getValue() instanceof FieldAccess) {
-      Expression vehicleExpr = args.getFirst().expression.getValue();
-      return isDirectRider(vehicle, vehicleExpr) || isJointRider(vehicle, vehicleExpr);
-    }
-    return false;
-  }
-
-  private boolean isDirectRider(UserField vehicle, Expression vehicleExpr) {
-    return vehicleExpr instanceof FieldAccess fa && fa.field.getValue() == vehicle;
-  }
-
-  private boolean isJointRider(UserField vehicle, Expression vehicleExpr) {
-    if (vehicleExpr instanceof MethodInvocation vehicleMethod) {
-      if (vehicleMethod.expression.getValue() instanceof FieldAccess) {
-        FieldAccess target = (FieldAccess) vehicleMethod.expression.getValue();
-        return target.field.getValue() == vehicle;
-      }
-    }
-    return false;
+    return codeGenerator.getRiders(vehicle);
   }
 
   @Override
   public Statement[] getDoStatementsForRemoveField(UserField field, Map<AbstractField, Statement> riders) {
-    List<Statement> doStatements = Lists.newLinkedList();
-    for (AbstractField rider : riders.keySet()) {
-      doStatements.add(SetUpMethodGenerator.createSetVehicleSceneStatement(rider));
-    }
-    doStatements.add(SetUpMethodGenerator.createSetVehicleNullStatement(field));
-    return ArrayUtilities.createArray(doStatements, Statement.class);
+    return codeGenerator.getDoStatementsForRemoveField(field, riders);
   }
 
   @Override
   public Statement[] getUndoStatementsForRemoveField(UserField field, Map<AbstractField, Statement> riders) {
-    Object instance = this.getInstanceInJavaVMForField(field);
-    AbstractField vehicleField = null;
-    if (instance instanceof Rider rider) {
-      SThing vehicleInstance = rider.getVehicle();
-      vehicleField = this.getFieldForInstanceInJavaVM(vehicleInstance);
-    }
-    Statement[] setupStatements = SetUpMethodGenerator.getSetupStatementsForInstance(false, instance, this.getActiveSceneInstance(), false);
-    Statement vehicleStatement = SetUpMethodGenerator.createSetVehicleFieldStatement(field, vehicleField);
-    Statement[] statements = new Statement[setupStatements.length + 1 + riders.size()];
-    statements[0] = vehicleStatement;
-    System.arraycopy(setupStatements, 0, statements, 1, setupStatements.length);
-
-    int i = setupStatements.length + 1;
-    for (Statement setVehicleStatement : riders.values()) {
-      statements[i++] = setVehicleStatement;
-    }
-    return statements;
+    return codeGenerator.getUndoStatementsForRemoveField(field, riders);
   }
 
   @Override
